final_project/requirements.txt
torch==2.5.1
torchvision==0.20.1
transformers==4.51.3
matplotlib==3.10.0
numpy==1.26.4
pillow==11.0.0
notebook==6.5.6